requirements.txt
Django==1.8.5
django-debug-toolbar==1.6
sqlparse==0.2.2
